% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{eta_splom}
\alias{eta_splom}
\title{Scatterplot matrix with correlations, densities and LOESS smoothers.}
\usage{
eta_splom(x, eta.sd = NULL, ...)
}
\description{
Scatterplot matrix with correlations, densities and LOESS smoothers.
}
