% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablefunctions.R
\name{generate.omega.covariance.matrix.HTML}
\alias{generate.omega.covariance.matrix.HTML}
\title{Generate an OMEGA covariance matrix table in HTML}
\usage{
generate.omega.covariance.matrix.HTML(
  nm_results,
  omega = nm_results$om_cov,
  digits = 3
)
}
\description{
Generate an OMEGA covariance matrix table in HTML
}
