% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{gof_ident}
\alias{gof_ident}
\title{Identity GOF plot}
\usage{
gof_ident(formula, data, groups = NULL, xylim = NULL, logxy = F,
  limxf = 0.05, legend = !is.null(groups), loess = "black", ...)
}
\description{
Identity GOF plot
}
