% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{gof_resid}
\alias{gof_resid}
\title{Residual GOF plot}
\usage{
gof_resid(formula, data, groups = NULL, xlim = NULL, ylim = NULL,
  logx = F, limxf = 0.05, legend = !is.null(groups), loess = T,
  ...)
}
\description{
Residual GOF plot
}
