% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_nm_output}
\alias{read_nm_output}
\title{Read NONMEM output.}
\usage{
read_nm_output(rundir = getwd(),
  runname = basename(normalizePath(rundir)),
  lst.file = file.path(rundir, sprintf("\%s.lst", runname)),
  ext.file = file.path(rundir, sprintf("\%s.ext", runname)),
  shk.file = file.path(rundir, sprintf("\%s.shk", runname)),
  phi.file = file.path(rundir, sprintf("\%s.phi", runname)),
  phm.file = file.path(rundir, sprintf("\%s.phm", runname)),
  cov.file = file.path(rundir, sprintf("\%s.cov", runname)),
  cor.file = file.path(rundir, sprintf("\%s.cor", runname)),
  bootstrap.file = file.path(rundir, "Bootstrap", "bootstrap_dir1",
  sprintf("raw_results_\%s.csv", runname)),
  meta = read_meta(file.path(rundir, "meta.yaml")),
  th_names = meta$namemap$theta, om_names = meta$namemap$omega,
  sg_names = meta$namemap$sigma, use.vcov = FALSE, ...)
}
\arguments{
\item{use.vcov}{Should the default OMEGA and SIGMA be on the variance/covariance scale instead of the SD/correlation scale?}
}
\description{
Read NONMEM output.
}
